improved display of turtle routing info, corrected some minor compilation issues
git-svn-id: http://svn.code.sf.net/p/retroshare/code/trunk@2248 b45a01b8-16f6-495d-af2f-9b41ad6348cc

+++ libretroshare/src/turtle/p3turtle.cc
@@ -1717,7 +1717,7 @@ static std::string printNumber(uint64_t num,bool hex=false)
 	{
 		char tmp[100] ;
 
-		if(num < (1<<32))
+		if(num < (((uint64_t)1)<<32))
 			sprintf(tmp,"%08x", uint32_t(num)) ;
 		else
 			sprintf(tmp,"%08x%08x", uint32_t(num >> 32),uint32_t(num & ( (1<<32)-1 ))) ;
